+++ src/lib/components/utilities/transition/Transition.svelte
@@ -2,7 +2,11 @@
     import type {IGlobalProperties} from "../../../types/global";
     import type {IHTML5Properties} from "../../../types/html5";
 
-    import {map_data_attributes, map_global_attributes} from "../../../util/attributes";
+    import {
+        format_css_variables,
+        map_data_attributes,
+        map_global_attributes,
+    } from "../../../util/attributes";
 
     type $$Events = {
         animationend: AnimationEvent;
@@ -15,6 +19,7 @@
 
         animation?: "clip" | "fade" | "scale" | "slide";
         delay?: number | string;
+        duration?: number | string;
         direction?: "bottom" | "left" | "right" | "top";
         variation?: "enter" | "exit";
     } & IHTML5Properties &
@@ -33,8 +38,11 @@
 
     export let animation: $$Props["animation"] = undefined;
     export let delay: $$Props["delay"] = undefined;
+    export let duration: $$Props["duration"] = undefined;
     export let direction: $$Props["direction"] = undefined;
     export let variation: $$Props["variation"] = undefined;
+
+    $: _variables = format_css_variables({delay, duration});
 </script>
 
 <div
@@ -44,7 +52,7 @@
     {...map_data_attributes({animation, direction, variation})}
     on:animationend
     on:animationstart
-    style={delay ? `${style}--delay:${delay};` : style}
+    style={_variables ? `${style}${_variables};` : style}
 >
     <slot />
 </div>